+++ cli/src/commands/workspace/root.rs
@@ -14,26 +14,62 @@
 
 use std::io::Write as _;
 
+use clap_complete::ArgValueCandidates;
 use jj_lib::file_util;
+use jj_lib::ref_name::WorkspaceNameBuf;
+use jj_lib::workspace_store::SimpleWorkspaceStore;
+use jj_lib::workspace_store::WorkspaceStore as _;
 use tracing::instrument;
 
 use crate::cli_util::CommandHelper;
 use crate::command_error::CommandError;
 use crate::command_error::user_error;
+use crate::complete;
 use crate::ui::Ui;
 
-/// Show the current workspace root directory
+/// Show the workspace root directory
 #[derive(clap::Args, Clone, Debug)]
-pub struct WorkspaceRootArgs {}
+pub struct WorkspaceRootArgs {
+    /// Name of the workspace (defaults to current)
+    #[arg(long, value_name = "NAME", add = ArgValueCandidates::new(complete::workspaces))]
+    name: Option<WorkspaceNameBuf>,
+}
 
 #[instrument(skip_all)]
 pub fn cmd_workspace_root(
     ui: &mut Ui,
     command: &CommandHelper,
-    _args: &WorkspaceRootArgs,
+    args: &WorkspaceRootArgs,
 ) -> Result<(), CommandError> {
-    let loader = command.workspace_loader()?;
-    let path_bytes = file_util::path_to_bytes(loader.workspace_root()).map_err(user_error)?;
+    let path = if let Some(ws_name) = &args.name {
+        let workspace_command = command.workspace_helper_no_snapshot(ui)?;
+        let workspace_store = SimpleWorkspaceStore::load(workspace_command.repo_path())?;
+
+        if workspace_command
+            .repo()
+            .view()
+            .wc_commit_ids()
+            .contains_key(ws_name)
+        {
+            workspace_store
+                .get_workspace_path(ws_name)?
+                .ok_or_else(|| {
+                    user_error(format!(
+                        "Workspace has no recorded path: {}",
+                        ws_name.as_symbol()
+                    ))
+                })?
+        } else {
+            return Err(user_error(format!(
+                "No such workspace: {}",
+                ws_name.as_symbol()
+            )));
+        }
+    } else {
+        command.workspace_loader()?.workspace_root().into()
+    };
+
+    let path_bytes = file_util::path_to_bytes(&path).map_err(user_error)?;
     ui.stdout().write_all(path_bytes)?;
     writeln!(ui.stdout())?;
     Ok(())