+++ lib/src/lib.rs
@@ -116,6 +116,7 @@ pub mod union_find;
 pub mod view;
 pub mod working_copy;
 pub mod workspace;
+pub mod workspace_store;
 
 #[cfg(test)]
 mod tests {